+++ packages/std/src/lib.rs
@@ -38,7 +38,10 @@ pub use crate::ibc::{
     IbcPacketReceiveMsg, IbcPacketTimeoutMsg, IbcReceiveResponse, IbcTimeout, IbcTimeoutBlock,
 };
 #[cfg(feature = "iterator")]
-pub use crate::iterator::{Order, Pair};
+#[allow(deprecated)]
+pub use crate::iterator::Pair;
+#[cfg(feature = "iterator")]
+pub use crate::iterator::{Order, Record};
 pub use crate::math::{
     Decimal, Decimal256, Decimal256RangeExceeded, DecimalRangeExceeded, Fraction, Uint128, Uint256,
     Uint512, Uint64,